Green the rest of CI: QA/JET, Downgrade, Spell Check, Runic, Julia-pre
Builds on the MTK 11 / Symbolics 7 migration to take the remaining red
checks to green.

QA (JET.test_package) — fixed all 13 static-analysis findings at the source
(verified clean on Julia 1.12 and lts):
- generate_{raw,central}_moment_eqs(sys::System, ...): extract get_iv /
  get_noise_eqs once and error early if missing, narrowing the
  Union{Nothing,...} the new MTK accessors return out of the inner call
  (removes the iv::Nothing / noise::Nothing no-method dead-ends).
- log_normal_closure / derivative_matching: bind `M` on all branches
  (M = Dict() in the raw-equations branch) so JET no longer reports
  "M may be undefined" / getindex(::Nothing, ...) on the central-only paths.
- gen_iter(n::Int, d::Int): annotate the args so NTuple{n,Int} and the
  Combinatorics.partitions iterate resolve to concrete types.
- get_moments_FSP: construct the moment dict as Dict{NTuple{N,Int},Float64}()
  (was Dict{NTuple{N,Float64}}(), an incomplete value-type with no zero-arg
  constructor and the wrong key element type).

Downgrade — raise stale compat floors so the minimum-version resolve no
longer selects an MTK 9.x that fails to precompile (NonlinearProblem has no
field tspan in MTKChainRulesCoreExt). The code now uses MTK 11-only API
(System, get_noise_eqs), so the old majors could never load anyway. Floors
raised to the minimum mutually-compatible set for Catalyst 16.2 / MTK 11.26:
Catalyst "16.2", ModelingToolkit "11.26", SciMLBase "3.21", Symbolics "7.28",
SymbolicUtils "4.35", Cumulants "1.0.6", DataStructures "0.19.5",
Distributions "0.25.127", DocStringExtensions "0.9.5", Latexify "0.16.10",
Combinatorics "1.0.2", TupleTools "1.6", JumpProcesses "9.29",
OrdinaryDiffEqTsit5 "2.0.1", SciMLTesting "1.2" (the arg-less run_tests()
in test/runtests.jl needs SciMLTesting >= 1.2; 1.0 lacks the `core` kwarg
default and errors at downgrade). Verified: full floor-pin resolve +
precompile + Core test (incl. latexify) on lts.

Spell Check — fix typos (occurence->occurrence, scenarion->scenario,
remaning->remaining, parameteric->parametric, dependendent->dependent);
ignore Jupyter notebook cell-id hashes ("id": "<hex>") in _typos.toml.

Runic — strip a trailing zero (65.595904620 -> 65.59590462) in
moment_ODE_tests.jl.

Julia pre (1.13) — two version-sensitive test comparisons:
- the log-normal closure_tests comparison still used
  isequal(Fixpoint(simplify)(...)); switch it to the same `symeq` numeric
  check used elsewhere (1.13's simplifier produces a different but
  mathematically-equal canonical form).
- the latexify closure test pinned the exact LaTeX string, but the
  single-fraction numerator's commuting factors print in a version-dependent
  order (μ_{1 0} μ_{1 2} μ_{1 1} on latest lts, μ_{1 0} μ_{1 1} μ_{1 2} on
  1.13, μ_{1 2} μ_{1 1} μ_{1 0} at the downgrade floor). Compare that form up
  to factor reordering by sorting whitespace-separated tokens (still pins the
  exact factor/term set and denominators), keeping the two-fraction forms as
  exact alternatives.

+++ src/symbolic.jl
@@ -1,4 +1,4 @@
-function gen_iter(n, d)
+function gen_iter(n::Int, d::Int)
     # based on https://twitter.com/evalparse/status/1107964924024635392
     iter = NTuple{n, Int}[]
     for x in partitions(d + n, n)
